u-boot: Fix implicit declaration compile error
A recent patch to ti-u-boot 2025.01 introduces a compile error on
master for am64xx-evm.  A patch has been sent to the mailing list
that fixes this, but it will take time before we will see that patch
available to us on master.  So we need to carry the patch here.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>