wic: sdimage-2part-efi: Add variable for adding bootloader options
We have a need for some platforms to add kernel options to the bootloader
during wic creation time.  Create a new variable and add that to the
sdimage-2part-efi.wks.in file so that a machine can override it if
needed.

Add an override for am64 to point the kernel at console=ttyS2.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>