ti-bsp: ti-img-encode-decode: Add Recipe for Building VXE-VXD Out-of-Tree Module
During 10.0 LTS bringup, the decision was made that it would be the last LTS
that carried the IMG VXE and VXD driver in tree. Driver is now hosted on
git.ti.com. This recipe is responsible for pulling in the driver,
building it, and packaging it within the J721e SDK.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>