conf: machine: j721s2/j784s4: Fix overlays for DS90UB954
The previous commit incorrectly added the k3-j721s2-evm-ub954.dtbo to
the KERNEL_DEVICETREE and it was missed during a quick review.

The correct way to add dtbs is to add a prefix string to the
KERNEL_DEVICETREE_PREFIX variable so that all files that match the
string will be picked up.  This allows for multiple kernels to be
swapped out that may have different DTBs in their respective trees and
not break a build.

The upstream testing on the kernel will automatically update the
KERNEL_DEVICETREE variable when the final kernel version is available.
That way the KERNEL_DEVICETREE always points to what is available in the
mainline.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>