conf: machine: j721e: Add BSP support for new out of tree module
The new out of tree module ti-vxe-vxd-driver is only for the 6.12 kernel
and later.  Since we support multiple BSPs we need to correctly wrap
this addition behind some additional logic that can make sure to not
include it for the 6.1 and 6.6 kernel builds where the module is still
in the tree.

Additionally, fix up the beaglebone-ai64 since it includes the j721e.inc
file and will need to clear this variable for its BSPs as well.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>