optee-os: drop custom handling of embedded TAs
The meta-arm layer, which is the provider of the base recipe for
optee-os, added [1] a separate optee-os-ta package for the embedded TAs
that are compiled as part of the OPTEE OS build process. However, the
TI overrides for optee-os include the embedded TAs inside the base
package, leaving the optee-os-ta package empty (and thus not
installable).

Align to the meta-arm behaviour and simply add a runtime dependency on
optee-os-ta for optee-os. While keeping backward compatibility, this
allows to install only the embedded TAs - reducing the required disk
space - on platforms on which the OPTEE OS is loaded from an external
boot image rather than being fetched from the root filesystem, like the
K3 SoCs.

[1] https://patchwork.yoctoproject.org/project/arm/patch/20230630112340.3314395-1-r.czerwinski@pengutronix.de/

Signed-off-by: Francesco Valla <francesco@valla.it>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>