bsp: tfa: add build flags for ti-falcon
In falcon mode, ATF jumps straight to the linux kernel which requires a
2MiB aligned load address. Hence this patch changes the ATF jump address
(PRELOADED_BL33_BASE) to a 2MiB aligned value 0x82000000 and the DTB
address (K3_HW_CONFIG_BASE) to 0x88000000.

These values are chosen to be consistent with the addresses used by
u-boot when booting in the normal flow.

Signed-off-by: Anshul Dalal <anshuld@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>