CLEANUP: :append needs a space
Usually :append is used to add items to a space separated list.  In
those cases the first character of the appended string should be a
space to make sure that the append is correct.

In this case, the space was embedded in the string and not right after
the ".  Simple fix is to move the space to the front to be consistent.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>