linux: Temporarily backport previous fitimage classes
A recent set of changes on oe-core has dramatically revamped the
fitimage generation code.  It is initially unclear how to fold those
change into our flow, so in the short term we will copy the previous
versions of the kernel bbclass files from a commit before [1] any
changes to the class files.

As soon as we figure it out, these files will be removed.  This just
help unblock anyone trying to build on master including meta-ti-bsp.

[1] https://git.openembedded.org/openembedded-core/commit/?id=fdb0a51598156f99aa91f7495d7eada92a459e97

Signed-off-by: Ryan Eatmon <reatmon@ti.com>