meta-beagle: linux: Add INSANE_SKIP debug-deps
We are seeing an error related to a check from oe-core wanting to make
sure that a package does not RDEPEND on a debug package.

ERROR: linux-bb.org-6.12.22+git-r0 do_package_qa: QA Issue: kernel-modules rdepends on kernel-module-g-dbgp-6.12.22-ti [debug-deps]

A bug has been filed:

https://bugzilla.yoctoproject.org/show_bug.cgi?id=15901

In the meantime, add debug-deps to the INSANE_SKIP list.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>