meta-ti/linux: Update for S changes
The kernel class operates a little differently.  The kernel is
downloaded and then moved to the "global" shared location and a symlink
is created to point to that location.  Since we cannot set S =
"${WORKDIR}/git" anymore, just choose a new name under UNPACKDIR.  The
value for BB_GIT_DEFAULT_DESTSUFFIX is the subdir name that is used by
default when fetching git repositories, so just use that name.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>