ti-ipc-rtos: Update to handle nested git repos
When specifying multiple git repos in the SRC_URI and you want the
"secondary" repositories to appear within the directory structure of the
primary repository, then we need to make sure that the destsuffix is
correct.  With the change to calculating the value of S to use
BB_GIT_DEFAULT_DESTSUFFIX, we need to make sure that the destsuffix is
based on that variable as well.

Also, update the paths for S to be within new source location.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>