trusted-firmware-a: Move back to tfa upstream except for am62lxx
This moves all of the platforms back to the main TFA repository, but
still points to the TI fork for the am62lxx platform while we continue
to work on upstreaming the needed patches.

With the the split into two repositories on two branches we need to also
split the LIC_FILES_CHECKSUM into two since the license.rst file is
slightly different between the two branches.  The upstream master
version is newer than the current am62l version.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>