u-boot-ti: Move UBOOT_VERSION into dynamic layer
The gitpkgv class we are using to set UBOOT_VERSION is located in
meta-openembedded.  We do not want to depend on meta-openembedded to
keep meta-ti-bsp as light as possible, and the naming of UBOOT_VERSION
is not a requirement, just a nice to have.  Dynamic layers allow us to
use the class if it is available in the build, but not require it.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>