linux-ti-staging-rt_6.12: Need finer control for BRANCH
The CICD flow needs to control the value for BRANCH in order to have the
recipe updated with the next SRCREV, but point to the CICD/next branch
for initial build/test.

With this split in linux-ti-staging-rt we cannot just have an override
for BRANCH in the local.conf because it trumps the new BRANCH_ARM32
system.

This patch creates an additional "default" variable BRANCH_ARM64 (or not
arm32), which can be controlled separately by the CICD flow.  Then BRANCH
simply is picking between the two and so can be controlled by overrides
in the recipe.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>