linux-ti-*: Fix support for building legacy HS platforms
The legacy HS platforms (am335x-hs-evm, am437x-hs-evm, and
am57xx-hs-evm) are not using the new fitImage flow, but the kernels all
include the kernel.bbclass which disallows the old fitImage flow.

Copy in the last version of the kernel.bbclass and name it
kernel_legacyhs.bbclass, and update the machine configs to defined a
variable to use as the KERNEL_BASE_CLASS to inherit.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>