conf: machine: add optee-os into am62lxx-evm builds
Newer silicon revision of the AM62L family are capable of reliably
booting with OP-TEE integrated in the bootflow with all bootmedia.
However to support as many DDR densities as possible we cannot use the
default 0x9e78000 location as with other K3 devices.

So add OP-TEE back into the default build but modify the BL32 base
address for TF-A so it will jump to the correct location.

Signed-off-by: Bryan Brattlof <bb@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>