ti-extras: Clean up
The TI_EXTRAS infrastructure needs a little tweaking to support turning
on multiple extra options at the same time.

1) Since TI_EXTRAS can be a space separated list and not a single fixed
value, we need to change how we detect when to include the TI_EXTRA
values.

2) The tie-jailhouse overrides were being a little too heavy handed.
All of the variables that are changed for any given extra feature needs
to use the proper variable override.  This was most important for the
COMPATIBLE_MACHINE settings.

3) Move the ti-extras out of specific machine configs and into the
ti-soc so that all machines can use the extras.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>