meta-ti-extras: dsptop: Simplify for DRA7xx only build
This recipe is already only compatible with the DRA7xx machine, so
no need to set PLATFORM based on the only possibility. This then
allows us to directly set the compile flags for the same. And now
that the package does not change based on the machine, remove the
PACKAGE_ARCH line.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>