meta-ti-extras: gdbc6x: Explicitly allow build for AM57x
This recipe builds and runs fine on both DRA7x and AM57x machines.
The COMPATIBLE_MACHINE is set to just DRA7x, but this was okay as AM57x
machines also set themselves as DRA7x family devices. The issue is then
we cannot then differentiate between packages that are actually DRA7x-
only and those that just use that family to mean both machine types. To
fix this when we mean both are allowed use "omap-a15" as done here.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>