machine: Add FIT_CONF_DEFAULT_DTB for legacy platforms
As of u-boot 2025.10, the fit image requires a value for the default
configuration.  The change [1] introduced the FIT_CONF_DEFAULT_DTB code
to the legacy fitimage class, but failed to set values for the legacy
platforms.  This led to us putting default = "" in the .its file, which
is now an error.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>