packagegroup-ti-core-initramfs: ensure TI_INITRAMFS_KERNEL_MODULES is set
TI_INITRAMFS_KERNEL_MODULES gets set only for TI platfroms and ends up
being undefined otherwise, which breaks parsing in yocto-check-layer
and/or in multi-BSP setups. Ensure the variable is defined in the
packagegroup, if not already set by the machine configuration.

Remove the previous default assignment since we put it in the wrong
location.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>