layer: add github MIRRORS definitions
We have some projects mirrored to GitHub currently. It makes sense to
add a MIRRORS entry for git.ti.com sources to fall back on in the event
git.ti.com is unreachable.

Placing this in the main BSP layer configuration file should be all that
is required for this and any other layers that depend this core layer to
fall back to GitHub.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>