k3r5: Use UBOOT_BINARY for R5 U-Boot Image
As R5 only builds one working image (tiboot3.bin), define this as the
"UBOOT_BINARY", even though it was technically build using the SPL build
pass in U-Boot. This allows some simplification and removes deploying of
a dummy file (u-boot-spl.bin) which is not a usable boot file. It was only
deployed as UBOOT_BINARY is not optional and so something needed to be set
here.

BeagleBone-AI64 already did something similar to this so the overrides
in that MACHINE conf can now be removed.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>