am62xx-evm-k3r5: Centralize the k3r5 settings
Since we are going to be effectively duplicating most of the what k3r5
does for the k3r5-falcon builds, we need to have the same settings.  But
we cannot just include the am62xx-evm-k3r5.conf as a starting point
since it sets somethings we do not want.

Solution is to pull all of the common settings out into an include files
and just include it in both flows.

This commit just does this for the existing k3r5 build so that the
upcoming k3r5-falcon change can use it.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>