u-boot: Update trusted-firmware-a locations
A recent change in meta-arm installs all of files into a
trusted-firmware-a directory.  We need that added to the paths for the
PACKAGECONFIG settings so that u-boot can compile.

Take this opportunity to also break all of the firmware paths into
separate variables in case we need to override one of them in the future.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>